Added some base level code for authentication, starting to add google auth
