+++ frontend/screens/LoginScreen.tsx
@@ -1,0 +1,10 @@
+import React, { useState, useContext } from 'react'
+import {
+    View,
+    StyleSheet,
+    TouchableWithoutFeedback,
+    Keyboard
+} from 'react-native'
+import Firebase from '../Firebase'
+import { TextInput, Button } from 'react-native'
+
